+++ src/hooks/poll/useOptimisticPoll.ts
@@ -1,0 +1,4 @@
+const useOptimisticPoll = () => {
+  const OptimsticPollUp = () => {};
+  const OptimsticPollDown = () => {};
+};
