+++ src/pages/MovieDetail/MovieDetailBeforeOpen.tsx
@@ -1,0 +1,203 @@
+import styled from 'styled-components';
+import { Txt } from '@/components/common/Txt';
+import { IMovieDetail } from '@/models/movie.model';
+import { IPollBox } from '@/models/poll.model';
+import { IStockMovieInfo } from '@/models/stock.model';
+import { FcAdvertising } from 'react-icons/fc';
+import VoteButton from '@/components/vote/voteButton';
+import { dateDifference } from '@/utils/format';
+import { Link } from 'react-router-dom';
+import colors from '@/constants/colors';
+import clockImg from '@/assets/images/movieDetail/clock.svg';
+import rightArrow from '@/assets/images/main/arrow.png';
+
+interface Props {
+  movieDetail: IMovieDetail;
+  pollBox: IPollBox;
+  stockMovieInfo: IStockMovieInfo;
+}
+
+export default function MovieDetailBeforeOpen({
+  movieDetail,
+  pollBox,
+  stockMovieInfo,
+}: Props) {
+  const handleUpVote = () => {
+    console.log('오른다 선택됨');
+  };
+
+  const handleDownVote = () => {
+    console.log('내린다 선택됨');
+  };
+
+  return (
+    <RightContainer>
+      <VotingNowTxt typography="Pretendard36bold" color="watcha">
+        Now voting
+      </VotingNowTxt>
+
+      <Txt typography="Pretendard20bold">
+        <FcAdvertising /> '{movieDetail.movieTitle}' 개봉 후&nbsp;
+        {stockMovieInfo.companyName}의 주가가 오를까요? 내릴까요?
+      </Txt>
+
+      <div className="movie-description">
+        <p>{movieDetail.movieDescription}</p>
+      </div>
+
+      <Divider />
+
+      <ScoreInfo typography="Pretendard24bold">
+        🎯 정답을 맞추면 10포인트 획득!
+      </ScoreInfo>
+
+      <VoteContainer>
+        <VoteButton onUpVote={handleUpVote} onDownVote={handleDownVote} />
+      </VoteContainer>
+
+      <div className="voting-status">
+        <div className="voting-status-left">
+          <Txt typography="Pretendard24bold" color="watcha">
+            {dateDifference(movieDetail.movieOpenDate)} 후 종료
+          </Txt>
+          <Txt typography="Pretendard24bold" color="white" className="divide">
+            {(pollBox.up + pollBox.down).toLocaleString()} 명 참여 중
+          </Txt>
+        </div>
+        <div className="voting-status-right">
+          <Link to="/movie-list">
+            <Txt
+              typography="Pretendard24regular"
+              color="black"
+              className="move"
+            >
+              다른 영화 예측
+            </Txt>
+          </Link>
+        </div>
+      </div>
+    </RightContainer>
+  );
+}
+
+// 오른쪽 섹션 (간단 설명 및 투표)
+const RightContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 10px;
+  flex: 1;
+  margin-left: 40px;
+  width: 100%;
+
+  .movie-description {
+    width: 588px;
+    max-height: 128px;
+    height: 100%;
+
+    overflow: scroll;
+    display: -webkit-box;
+    -webkit-line-clamp: 4;
+    -webkit-box-orient: vertical;
+    p {
+      font-size: 16px;
+      color: ${colors.greyscale10};
+      line-height: 1.6;
+      text-align: center;
+    }
+  }
+
+  // 투명 스크롤바
+  .movie-description::-webkit-scrollbar {
+    display: none;
+  }
+
+  .voting-status {
+    width: 750px;
+    height: 64px;
+    border-radius: 20px;
+    background-color: ${colors.watcha3};
+
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    padding: 0 10px 0 10px;
+
+    .voting-status-left {
+      display: flex;
+      align-items: center;
+
+      .divide {
+        display: flex;
+      }
+
+      .divide::before {
+        content: '/';
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        width: 35px;
+        height: 35px;
+      }
+    }
+    .voting-status-left::before {
+      content: '';
+      display: block;
+      width: 35px;
+      height: 35px;
+      background-image: url(${clockImg});
+      background-size: cover;
+      background-position: center;
+    }
+
+    .voting-status-right {
+      display: flex;
+      align-items: center;
+
+      .move {
+        display: flex;
+        align-items: center;
+      }
+      .move::after {
+        content: '';
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        width: 25px;
+        height: 25px;
+        background-image: url(${rightArrow});
+        background-size: cover;
+        background-position: center;
+      }
+    }
+  }
+`;
+
+const VotingNowTxt = styled(Txt)`
+  border: 1px solid ${colors.watcha};
+  border-radius: 16px;
+  padding: 4px 10px;
+`;
+
+const MovieDescription = styled(Txt)`
+  line-height: 1.6;
+  display: -webkit-box;
+  -webkit-line-clamp: 4;
+  -webkit-box-orient: vertical;
+`;
+
+const Divider = styled.div`
+  width: 100%;
+  height: 1px;
+  background-color: ${colors.greyscale8};
+  margin: 10px 0;
+`;
+
+const ScoreInfo = styled(Txt)`
+  margin-bottom: 20px;
+`;
+
+const VoteContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+`;
